+++ src/WebHaven.TelegramBot/Bot/MessageHandlers/AddFeedCommandHandler.cs
@@ -2,7 +2,7 @@
 using Telegram.Bot.Types.ReplyMarkups;
 using WebHaven.TelegramBot.Bot.UserLogic;
 
-namespace WebHaven.TelegramBot.Bot.Handlers.MessageHandlers;
+namespace WebHaven.TelegramBot.Bot.MessageHandlers;
 
 public class AddFeedCommandHandler(ITelegramBotClient bot, UserRepository userRepo) : IMessageHandler<AddFeedCommand>
 {
