+++ src/WebHaven.TelegramBot/Bot/MessageHandlers/GettingFeedMenuHandler.cs
@@ -5,7 +5,7 @@ using Telegram.Bot.Types.ReplyMarkups;
 using WebHaven.TelegramBot.Bot.UserLogic;
 using WebHaven.TelegramBot.Feeds;
 
-namespace WebHaven.TelegramBot.Bot.Handlers.MessageHandlers;
+namespace WebHaven.TelegramBot.Bot.MessageHandlers;
 
 public class GettingFeedMenuHandler(
         ITelegramBotClient bot,
